+++ OptionItem.c
@@ -62,7 +62,7 @@ static void NumberItem_display(const Object* cast, RichString* out) {
    } else {
       written = xSnprintf(buffer, sizeof(buffer), "%d", NumberItem_get(this));
    }
-   RichString_appendnAscii(out, CRT_colors[CHECK_MARK], buffer, written);
+   RichString_appendnAscii(out, CRT_colors[CHECK_MARK], buffer, (unsigned int)written);
    RichString_appendAscii(out, CRT_colors[CHECK_BOX], "]");
    for (int i = written; i < 5; i++) {
       RichString_appendAscii(out, CRT_colors[CHECK_BOX], " ");
