+++ TasksMeter.c
@@ -46,21 +46,21 @@ static void TasksMeter_display(const Object* cast, RichString* out) {
    int len;
 
    len = xSnprintf(buffer, sizeof(buffer), "%d", (int)this->values[2]);
-   RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, len);
+   RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, (unsigned int)len);
 
    RichString_appendAscii(out, settings->hideUserlandThreads ? CRT_colors[METER_SHADOW] : CRT_colors[METER_TEXT], ", ");
    len = xSnprintf(buffer, sizeof(buffer), "%d", (int)this->values[1]);
-   RichString_appendnAscii(out, settings->hideUserlandThreads ? CRT_colors[METER_SHADOW] : CRT_colors[TASKS_RUNNING], buffer, len);
+   RichString_appendnAscii(out, settings->hideUserlandThreads ? CRT_colors[METER_SHADOW] : CRT_colors[TASKS_RUNNING], buffer, (unsigned int)len);
    RichString_appendAscii(out, settings->hideUserlandThreads ? CRT_colors[METER_SHADOW] : CRT_colors[METER_TEXT], " thr");
 
    RichString_appendAscii(out, settings->hideKernelThreads ? CRT_colors[METER_SHADOW] : CRT_colors[METER_TEXT], ", ");
    len = xSnprintf(buffer, sizeof(buffer), "%d", (int)this->values[0]);
-   RichString_appendnAscii(out, settings->hideKernelThreads ? CRT_colors[METER_SHADOW] : CRT_colors[TASKS_RUNNING], buffer, len);
+   RichString_appendnAscii(out, settings->hideKernelThreads ? CRT_colors[METER_SHADOW] : CRT_colors[TASKS_RUNNING], buffer, (unsigned int)len);
    RichString_appendAscii(out, settings->hideKernelThreads ? CRT_colors[METER_SHADOW] : CRT_colors[METER_TEXT], " kthr");
 
    RichString_appendAscii(out, CRT_colors[METER_TEXT], "; ");
    len = xSnprintf(buffer, sizeof(buffer), "%d", (int)this->values[3]);
-   RichString_appendnAscii(out, CRT_colors[TASKS_RUNNING], buffer, len);
+   RichString_appendnAscii(out, CRT_colors[TASKS_RUNNING], buffer, (unsigned int)len);
    RichString_appendAscii(out, CRT_colors[METER_TEXT], " running");
 }
